moved class `CppDict` from module `dictIO.dict` into `dictIO.cppDict` and marked both the module and the `CppDict` class as deprecated.
